api/requirements.txt
# FastAPI et serveur ASGI
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Sécurité et authentification
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email]==2.6.1
pydantic-settings==2.1.0

# Base de données
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Sécurité avancée
slowapi==0.1.9
python-dotenv==1.0.1
argon2-cffi==23.1.0

# Validation et sanitisation
bleach==6.1.0
email-validator==2.1.0.post1

# CORS et headers sécurisés
secure==0.3.0

# Logging sécurisé
structlog==24.1.0

# Tests
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
